Make supervisor cd to '/data' on daemon startup.
Some programs don't respect the 'directory' setting in their individual
'.ini' file and use the same directory supervisor is started in. When
unset as before, this was '/'. Changing to '/data' makes sure that any
temp files or other clutter that programs create is sure to be in a
location we can manage and expect.