Change socketi/pid location to use "%(here)s".
Now we trick Supervisor into storing it's socket and pid files in a
dynamic directory, (/run/supervisor/$HOSTNAME). The supervisor.conf file
doesn't allow use of %(host_node_name)s when defining a pid or
socketfile. But we want that, so we use $(here)s which uses the location
of the configuration file itself.